requirements.txt
Django==1.6.1
psycopg2==2.5.1

Fabric==1.8.2
django-kronos==0.4

# for scraping new schedule
requests==2.3.0
beautifulsoup4==4.3.2